renamed Pack container to PackingOverview, this now will be the parent to packing and reminders
